move `Address` and `Page` to use u64

+++ src/kernel/src/arch/x64/registers/control/cr2.rs
@@ -5,7 +5,7 @@ pub struct CR2;
 impl CR2 {
     /// Read the current page fault linear address from the CR2 register.
     pub fn read() -> Address<Virtual> {
-        let value: usize;
+        let value: u64;
 
         unsafe {
             core::arch::asm!("mov {}, cr2", out(reg) value, options(nomem, nostack, preserves_flags));
